tcp port are always stored in an unsigned short
git-svn-id: https://svn.code.sf.net/p/gopherus/code/trunk@17 767d67df-c2ba-418b-950d-eab06a56d38d

+++ history.c
@@ -43,7 +43,7 @@ void history_back(struct historytype **history) {
 
 
 /* adds a new node to the history list. Returns 0 on success, non-zero otherwise. */
-int history_add(struct historytype **history, char protocol, char *host, unsigned int port, char itemtype, char *selector) {
+int history_add(struct historytype **history, char protocol, char *host, unsigned short port, char itemtype, char *selector) {
   struct historytype *result;
   int tmplen;
